feat: address review comments
- bunch of documentation
- better variable names
- moved some more non-UI related code out of UI code
- log offending images to sentry in case of unexpected failure
